software_design_project/Code/contenedorOCR/requirements.txt
pymongo
tesseract
pillow
pytesseract
regex
glob2
watchdog
